Split processor files out
