+++ src/download/main.py
@@ -1,3 +1,4 @@
+import argparse
 from datetime import datetime
 import urllib3
 from typing import List
@@ -8,7 +9,7 @@ urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
         
 
 
-def download_meeting_files(start_year: int, start_month: int, end_year: int, end_month: int, meeting_filter: List[str] = None, file_filter: List[str] = None):
+def download_meeting_files(start_year: int, start_month: int, end_year: int, end_month: int, meeting_filter: List[str] = None, file_filter: List[str] = None, download_location: str = "./data/meetings/downloaded", log_file_folder: str = ".logs"):
     script_start_time = datetime.now()
     print(f"Fetching meeting minutes from {start_year}-{start_month} to {end_year}-{end_month}...")
     if meeting_filter:
@@ -19,7 +20,7 @@ def download_meeting_files(start_year: int, start_month: int, end_year: int, end
     start_date = datetime(start_year, start_month, 1)
     end_date = datetime(end_year, end_month, 1) # Using first date will still include the entire month
     current_date = start_date
-    downloader = MeetingFilesDownloader(destination_folder = "./data/meetings/downloaded", log_file_folder=".logs", log_file_name=f"{script_start_time.strftime('%Y-%m-%d_%H-%M-%S')}_failed_downloads.txt")
+    downloader = MeetingFilesDownloader(destination_folder = download_location, log_file_folder=log_file_folder, log_file_name=f"{script_start_time.strftime('%Y-%m-%d_%H-%M-%S')}_failed_downloads.txt")
     
 
     while current_date <= end_date:
@@ -34,3 +35,26 @@ def _add_month(date: datetime) -> datetime:
         return date.replace(year=date.year + 1, month=1)
     else:
         return date.replace(month=date.month + 1)
+
+if __name__ == "__main__":
+    # Parse command-line arguments
+    parser = argparse.ArgumentParser(description="Download council meeting minutes PDFs.")
+    parser.add_argument("--start-year", type=int, default=2014, help="Start year (e.g., 2023)")
+    parser.add_argument("--start-month", type=int, default=1, help="Start month (1-12)")
+    parser.add_argument("--end-year", type=int, default=datetime.now().year, help="End year (e.g., 2024)")
+    parser.add_argument("--end-month", type=int, default=datetime.now().month, help="End month (1-12)")
+    parser.add_argument("--meeting-filter", nargs='+', type=str, default=None, help="Filter meetings by names (case insensitive, e.g., 'council budget')")
+    parser.add_argument("--file-filter", nargs='*', type=str, default=["agenda", "minutes"], help="Filter files by names (case insensitive, e.g., 'agenda minutes'). Use '--file-filter' with no arguments to disable filtering.")
+    args = parser.parse_args()
+
+    if not args.file_filter:  # If file-filter is provided with no arguments
+        args.file_filter = None
+        
+    download_meeting_files(
+        start_year=args.start_year,
+        start_month=args.start_month,
+        end_year=args.end_year,
+        end_month=args.end_month,
+        meeting_filter=args.meeting_filter,
+        file_filter=args.file_filter,
+    )
